requirements.txt

flask
flask_cors
openai-whisper
torch
soundfile
